requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
trafilatura>=1.7.0
newspaper3k>=0.2.8
markdownify>=0.12.0
spacy>=3.7.1
textblob>=0.18.0.post0
tldextract>=5.1.2
python-dotenv>=1.0.1
openai>=1.30.0
rich>=13.7.1
typer>=0.12.3
pydantic>=2.7.0
